+++ examples/navbar/src/pages/navbar.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useRef, useMemo, Fragment } from 'react'
-import { useScrollPosition } from '@n8tb1t/use-scroll-position'
+import { useScrollPosition } from '../../../../src/'
 
 import NavLinks from '../components/nav-links'
 import FakeData from '../components/fake-data'
@@ -9,15 +9,15 @@ import { Navbar, Content, Footer } from '../styles'
 export default () => {
   const [hideOnScroll, setHideOnScroll] = useState(true)
   const rendersCount = useRef(0)
+  const refContainer = useRef(null)
 
   useScrollPosition(
+    refContainer,
     ({ prevPos, currPos }) => {
       const isShow = currPos.y > prevPos.y
       if (isShow !== hideOnScroll) setHideOnScroll(isShow)
     },
     [hideOnScroll],
-    null,
-    false,
     300
   )
 
@@ -27,7 +27,7 @@ export default () => {
         <Navbar show={hideOnScroll}>
           <NavLinks />
         </Navbar>
-        <Content mt="45">
+        <Content mt="45" ref={refContainer}>
           <FakeData />
         </Content>
         <Footer>